% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCAPI-package.R
\docType{package}
\name{RCAPI-package}
\alias{RCAPI}
\alias{RCAPI-package}
\title{RCAPI: R Internal C API Usage}
\description{
RCAPI is an example package for R's internal C API usage.
}
\author{
\strong{Maintainer}: Jordi Rosell \email{jroselln@gmail.com} (\href{https://orcid.org/0000-0002-4349-1458}{ORCID})

}
\keyword{internal}
